Extracted Blog List Item into it's own component
